+++ inc/minishell.h
@@ -6,7 +6,7 @@
 /*   By: lscheupl <lscheupl@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/29 11:02:04 by linux             #+#    #+#             */
-/*   Updated: 2025/01/31 16:42:33 by lscheupl         ###   ########.fr       */
+/*   Updated: 2025/01/31 19:14:50 by lscheupl         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -45,7 +45,7 @@ int    exec_general(char *input, t_ast *root, t_ms *ms);
 char *ft_strndup(const char *s, size_t n);
 void    ft_exec_builtin(t_node_cmd *node,t_ms *ms);
 char	*ft_find_path(char *expanded, t_ms *ms, char **args);
-char	*ft_expander(char *input, int start, int end, t_ms *ms);
+char	*ft_expander(char *to_be_expanded, t_ms *ms);
 char	*ft_isolate_first_word(char *expanded);
 char	*ft_check_access(char **tab_path, char *cmd);
 char	**ft_isolate_path(t_ms *ms);
